bugfix: bugfix: corrigido campo numérico do preço da entidade Produto

+++ src/main/java/com/senai/gestaoloja/dto/request/ProdutoRequest.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record ProdutoRequest(
         // colocar validações aqui
         @NotNull(message = "O nome deve ser informado.")
@@ -15,7 +17,7 @@ public record ProdutoRequest(
         String descricao,
         @NotNull(message = "O preço deve ser informado.")
         @Positive(message = "O preço deve ser maior que zero.")
-        Double preco,
+        BigDecimal preco,
 
         @NotNull(message = "A quantidade deve ser informado.")
         @Positive(message = "A quantidade deve ser maior que zero.")
