+++ src/main/java/com/senai/gestaoloja/dto/response/ProdutoResponse.java
@@ -1,10 +1,12 @@
 package com.senai.gestaoloja.dto.response;
 
+import java.math.BigDecimal;
+
 public record ProdutoResponse(
         Long id,
         String nome,
         String descricao,
-        Double preco,
+        BigDecimal preco,
         Integer quantidade
 ) {
 }
